requirements.txt
opencv
pytorch
numpy
